JWT Feature implementation.

+++ src/main/java/com/charmedteeth/currencyrate/model/RateConvertRequest.java
@@ -6,8 +6,8 @@ import lombok.Data;
 public class RateConvertRequest {
     private String baseCurr;
     private String qouteCurr;
-    private double amount;
-
+    private long amount;
+  
     public String getBaseCurr() {
         return baseCurr;
     }
@@ -28,7 +28,7 @@ public class RateConvertRequest {
         return amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(long amount) {
         this.amount = amount;
     }
